Updated loss functions
- Adversarial loss now reflects the Semantic Pyramic paper, with loss calculated as MSE
- Implemented co-normalization of real and reconstructed feature maps before loss is calculated, as described in the paper
- Added L1 norm to reconstruction and diversity losses, as described in the papers

+++ source/slideflow/gan/semantic/model.py
@@ -47,13 +47,13 @@ def create_generator(
 	noise_input = tf.keras.layers.Input((z_dim,), name='noise_input')
 	c = tf.keras.layers.Input((n_classes,), dtype=tf.bool, name='class_input')
 	input_layers = [noise_input, c, feature_tensors['image'], feature_tensors['image_vgg16']]
-	features_with_pool = [feature_tensors['fc8'],
-					 feature_tensors['fc7'],
-					 tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv0']),
-					 tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv1']),
-					 tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv2']),
-					 tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv3']),
-					 tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv4']),
+	features_with_pool = [tf.cast(feature_tensors['fc8'], dtype=tf.float32),
+					 tf.cast(feature_tensors['fc7'], dtype=tf.float32),
+					 tf.cast(tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv0']), dtype=tf.float32),
+					 tf.cast(tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv1']), dtype=tf.float32),
+					 tf.cast(tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv2']), dtype=tf.float32),
+					 tf.cast(tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv3']), dtype=tf.float32),
+					 tf.cast(tf.keras.layers.MaxPool2D((2,2))(feature_tensors['conv4']), dtype=tf.float32),
 	]
 	#reconstructed_features = []
 	x = noise_input
